requirements.txt
django
django-ckeditor
djangorestframework
pillow